youtube_video_finder/requirements.txt
google-api-python-client==2.118.0
youtube-transcript-api==0.6.2
anthropic==0.40.0
python-dotenv==1.0.1
